requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.3
gspread==5.10.0
oauth2client==4.1.3
sendgrid==6.10.0
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.8
python-multipart==0.0.6
omnidimension==1.0.0
